split code in BOA packages
